requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
twilio==7.16.4
python-dotenv==1.0.0
APScheduler==3.10.4